Fix logging setup
