+++ fotobot/exif/exifworker.py
