+++ fotobot/exif/styles.py
