+++ fotobot/services/helper.py
